Moving network service related tests
