+++ android-agent/src/test/java/io/opentelemetry/android/internal/services/network/detector/NetworkDetectorTest.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.android.instrumentation.network;
+package io.opentelemetry.android.internal.services.network.detector;
 
 import static org.junit.Assert.assertTrue;
